add a lot of endpoints

+++ server/src/models/note.sql.js
@@ -3,19 +3,19 @@ const { QueryTypes } = require("sequelize");
 
 // CREATE note
 //     noteTitle
-//     dataUrl
+//     dataId
 //     createdAt
 async function insertNote(info) {
   try {
     await sequelize.query(
-      `INSERT INTO Note(noteTitle, dataUrl, createdAt) values ('${info.noteTitle}', '${info.dataUrl}', '${info.createdAt}')`,
+      `INSERT INTO Note(noteTitle, dataId, createdAt) values ('${info.noteTitle}', '${info.dataId}', '${info.createdAt}')`,
       {
         type: QueryTypes.INSERT,
       }
     );
 
     newNote = await sequelize.query(
-      `SELECT * FROM Note WHERE noteTitle = '${info.noteTitle}' AND dataUrl = '${info.dataUrl}' AND createdAt = '${info.createdAt}')`,
+      `SELECT * FROM Note WHERE noteTitle = '${info.noteTitle}' AND dataId = '${info.dataId}' AND createdAt = '${info.createdAt}')`,
       {
         type: QueryTypes.SELECT,
       }
@@ -27,6 +27,8 @@ async function insertNote(info) {
         type: QueryTypes.INSERT,
       }
     );
+
+    await alterNoteCategories("INSERT", newNote.noteId, info.categories);
     return newNote;
   } catch (e) {
     console.error(e);
@@ -62,26 +64,29 @@ async function selectNoteByNoteId(noteId) {
   }
 }
 
-
-
 async function transferOwnership(noteId, oldOwnerId, newOwnerId) {
   try {
-  
-
-
-    await sequelize.query(
-      `DELETE FROM NoteAccess WHERE noteId = ${noteId} AND userId = ${oldOwnerId}`
-    ),
-      {
-        type: QueryTypes.DELETE,
-      };
+    oldAccess = await sequelize.query(
+      `SELECT * FROM NoteAccess WHERE noteId = ${noteId} AND userId= ${oldOwnerId} AND accessStatus= 'owner'`,
+      { type: QueryTypes.SELECT }
+    );
+    if (oldAccess.length < 1) {
+      throw new Error("You are not the owner, you cannot transfer ownnership!");
+    }
 
     await sequelize.query(
-      `UPDATE NoteAccess SET accessStatus = "owner" WHERE noteId = ${noteId} AND userId = ${newOwnerId}`
-    ),
-      {
-        type: QueryTypes.UPDATE,
-      };
+      `DELETE FROM NoteAccess WHERE noteId = ${noteId} AND userId = ${oldOwnerId}`,
+      { type: QueryTypes.DELETE }
+    );
+    // This UPDATE will only be called when the newOwnerId is valid
+    if (newOwnerId) {
+      await sequelize.query(
+        `UPDATE NoteAccess SET accessStatus = "owner" WHERE noteId = ${noteId} AND userId = ${newOwnerId}`
+      ),
+        {
+          type: QueryTypes.UPDATE,
+        };
+    }
   } catch (e) {
     throw new Error(e.message);
   }
@@ -101,10 +106,98 @@ async function selectAllAccessorsByNoteId(noteId) {
   }
 }
 
+async function alterNoteCommunity(command, commId, noteId) {
+  try {
+    if (command === "INSERT") {
+      return await sequelize.query(
+        `INSERT CommunityNote(communityId, noteId) VALUES ('${commId}', '${noteId}')`,
+        { type: QueryTypes.INSERT }
+      );
+    } else if (command === "DELETE") {
+      return await sequelize.query(
+        `DELETE FROM CommunityNote WHERE communityId = '${commId}' AND noteid = '${noteId}')`,
+        { type: QueryTypes.DELETE }
+      );
+    }
+  } catch (e) {
+    throw new Error(e.message);
+  }
+}
+
+async function alterNoteCategories(command, noteId, categories) {
+  try {
+    if (command === "INSERT") {
+      const valueList = categories.map((cate) => {
+        return `('${cate}', '${noteId}')`;
+      });
+      const arg = valueList.join(", ");
+
+      console.log(arg);
+      return await sequelize.query(
+        `INSERT NoteCategory(categoryName, noteId) VALUES ${arg}`,
+        { type: QueryTypes.INSERT }
+      );
+    } else if (command === "DELETE") {
+      for (let i = 0; i < categories.length; i++) {
+        const category = categories[i];
+        await sequelize.query(
+          `DELETE FROM NoteCategory WHERE categoryName = '${category}' AND noteId = '${noteId}'`,
+          { type: QueryTypes.INSERT }
+        );
+      }
+    }
+  } catch (e) {
+    throw new Error(e.message);
+  }
+}
+
+
+async function selectNoteAccessByNoteIdAndUserId(noteId, userId) {
+  try {
+    let data = await sequelize.query(
+      `SELECT * FROM NoteAccess WHERE noteId = '${noteId}' AND userId = '${userId}'`,
+      {
+        type: QueryTypes.SELECT,
+      }
+    );
+    return data[0];
+  } catch (e) {
+    throw new Error(e.message);
+  }
+}
+
+
+async function alterNoteAccess(command, noteId, userId, accessStatus) {
+  try {
+    if (command === "INSERT") {
+      return await sequelize.query(
+        `INSERT NoteAccess(noteId, userId, accessStatus) VALUES ('${noteId}', '${userId}', '${accessStatus}')`,
+        { type: QueryTypes.INSERT }
+      );
+    } else if (command === "DELETE") {
+      await sequelize.query(
+        `DELETE FROM NoteAccess WHERE noteId = '${noteId}', userId = '${userId}', accessStatus = '${accessStatus}'`,
+        { type: QueryTypes.DELETE }
+      );
+    } else if (command === "UPDATE") {
+      await sequelize.query(
+        `UPDATE NoteAccess SET accessStatus = '${accessStatus}' WHERE noteId = '${noteId}', userId = '${userId}'`,
+        { type: QueryTypes.UPDATE }
+      );
+    }
+  } catch (e) {
+    throw new Error(e.message);
+  }
+}
+
 module.exports = {
   insertNote,
   selectNotesByUserId,
   selectAllAccessorsByNoteId,
   transferOwnership,
   selectNoteByNoteId,
+  alterNoteCommunity,
+  alterNoteCategories,
+  alterNoteAccess,
+  selectNoteAccessByNoteIdAndUserId
 };
